fix(PyFr): Refactor PyFr code base to share as much as possible components
