+++ src/panels/run/RunCluster.js
@@ -6,7 +6,7 @@ import style    from 'HPCCloudStyle/ItemEditor.mcss';
 import theme    from 'HPCCloudStyle/Theme.mcss';
 
 export default React.createClass({
-  displayName: 'run/cluster',
+  displayName: 'panels/run/RunCluster',
 
   propTypes: {
     contents: React.PropTypes.object,
@@ -28,20 +28,21 @@ export default React.createClass({
   updateState() {
     this.setState({ busy: true });
     client.listClusterProfiles()
-      .then(resp => {
-        this.setState({
-          profiles: resp.data,
-          profile: resp.data[0],
-          busy: false,
+      .then(
+        resp => {
+          this.setState({
+            profiles: resp.data,
+            profile: resp.data[0],
+            busy: false,
+          });
+          if (this.props.onChange) {
+            this.props.onChange('profile', resp.data[0]._id, 'Traditional');
+          }
+        },
+        err => {
+          console.log('Error: Sim/RunCluster', err);
+          this.setState({ busy: false });
         });
-        if (this.props.onChange) {
-          this.props.onChange('profile', resp.data[0]._id, 'Traditional');
-        }
-      })
-      .catch(err => {
-        console.log('Error: Sim/RunCluster', err);
-        this.setState({ busy: false });
-      });
   },
 
   dataChange(event) {
