+++ src/workflows/index.js
@@ -1,6 +1,6 @@
 import CodeSaturn   from './code-saturn';
-import PyFr         from './pyfr';
-import PyFrExec     from './pyfr-exec';
+import PyFr         from './pyfr/pyfr-simput';
+import PyFrExec     from './pyfr/pyfr-exec';
 import Visualizer   from './visualizer';
 
 export default {
