+++ src/workflows/pyfr/pyfr-exec/logo.jpg
